requirements.txt
## Core backend
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.1
sqlalchemy==2.0.23
asyncpg==0.29.0
aioredis==2.0.1
web3==6.11.0
solana==0.30.0
solders==0.18.0

## Data/ML
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
transformers==4.36.2
torch==2.9.0
torchvision==0.24.0
xgboost==2.0.1
lightgbm==4.1.0
tensorflow==2.15.0

## RL
stable-baselines3==2.0.0
gym==0.26.2

## MLOps/Monitoring
mlflow==2.10.1
prometheus-client==0.19.0
grafana-api==1.0.4

## SDKs
coingecko-sdk==1.0.0

## Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pyyaml==6.0.1
asyncio==3.4.3
